DEMRUM-445: move helper methods into library module ../utils/iOSdSYMUtils.ts
